Updated Chat by adding exit feature
